backend/pyproject.toml
[project]
name = "job-recommender-backend"
version = "0.1.0"
description = "AI-powered job recommendation system backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.0",
    "email-validator>=2.1.0",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "sentence-transformers>=3.3.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.9.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "scikit-learn>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
only-packages = false

[tool.hatch.build.targets.wheel]
only-include = ["main.py", "schemas.py", "models", "services"]
